Allow versioning of extracted resources
Adds a new LiveDestination, which - for each extraction - writes the
added and removed triples and the Wikipedia page modification timestamp
to a database table (DBPEDIALIVE_REVISIONS).

Also adds a few library classes for working with revisions and
retrieving prior versions of resources for given timestamps:

org.dbpedia.extraction.timemachine:

- TimeMachine
- RevisionHistory
- Revision

In order to use these extensions, the `revisions.sql` file needs to be
loaded into your database management system. In `live.ini`, set the
`destinations.revision.enable` option to `true`.

Note: The LiveDestination#write signature was changed and now receives
and additional `timestamp: Long` argument.

Note: It may be advisable to not run the RevisionDestination and
PublisherDiffDestination in parallel - they both perform similar, but
not equivalent, tasks and could be rather resource heavy.